+++ mobile/src/main/kotlin/com/fesskiev/mediacenter/ui/main/MainContract.kt
@@ -37,6 +37,6 @@ interface MainContract {
 
         fun changePitchShift(value: Int)
 
-        fun changeTempo(value: Int)
+        fun changeTempo(value: Double)
     }
 }
